refactor(frontend): simplify session operation helpers
