+++ packages/frontend/src/state/operations/agent-orchestrator/handlers/session-actions-send.test.ts
@@ -1,12 +1,7 @@
 import { describe, expect, test } from "bun:test";
 import { OpencodeSdkAdapter } from "@openducktor/adapters-opencode-sdk";
 import { MANUAL_SESSION_COMPACTION_SLASH_COMMAND } from "@openducktor/contracts";
-import type {
-  AcceptedAgentUserMessage,
-  AgentEnginePort,
-  AgentUserMessagePart,
-} from "@openducktor/core";
-import { serializeAgentUserMessagePartsToText } from "@openducktor/core";
+import type { AcceptedAgentUserMessage } from "@openducktor/core";
 import { agentSessionIdentityKey } from "@/lib/agent-session-identity";
 import { getAgentSession, replaceAgentSession } from "@/state/agent-session-collection";
 import {
@@ -27,30 +22,7 @@ import {
   getSession,
 } from "./session-actions.test-helpers";
 import { createOpenCodeAgentEngineTestAdapter } from "./opencode-agent-engine.test-support";
-
-const acceptedUserMessage = (
-  input: Pick<
-    Parameters<AgentEnginePort["sendUserMessage"]>[0],
-    "externalSessionId" | "model" | "parts"
-  >,
-  messageId = "accepted-user-message",
-): AcceptedAgentUserMessage => {
-  // SAFETY: AgentEnginePort uses the same core message-part variants; generated optional fields differ only in declaration form.
-  const parts = input.parts as AgentUserMessagePart[];
-  const event: AcceptedAgentUserMessage = {
-    type: "user_message",
-    externalSessionId: input.externalSessionId,
-    timestamp: "2026-02-22T08:00:01.000Z",
-    messageId,
-    message: serializeAgentUserMessagePartsToText(parts),
-    parts: [],
-    state: "read",
-  };
-  if (input.model) {
-    event.model = input.model;
-  }
-  return event;
-};
+import { acceptedUserMessage } from "./session-actions-send.test-support";
 
 describe("agent-orchestrator/handlers/session-actions send", () => {
   test("routes a normalized workflow control without loading runtime policy settings", async () => {
